Made templateId and field version id a int rather than a long for Java.

+++ examples/java/uk/co/real_logic/sbe/examples/SbeExample.java
@@ -59,7 +59,7 @@ public class SbeExample
         MESSAGE_HEADER
             .reset(directBuffer, bufferOffset, messageTemplateVersion)
             .blockLength(CAR.blockLength())
-            .templateId((int)CAR.templateId())
+            .templateId(CAR.templateId())
             .version((short)CAR.templateVersion());
 
         bufferOffset += MESSAGE_HEADER.size();
